auth

+++ app/Http/Controllers/MessageController.php
@@ -3,11 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Models\Message;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 
 class MessageController extends Controller
 {
+    use AuthorizesRequests;
     public function index()
     {
 
@@ -26,9 +28,8 @@ class MessageController extends Controller
 
         ]);
 
-        Message::create([
-            'message' => $validated['message'],
-        ]);
+        auth()->user()->messages()->create($validated);
+
 
         return redirect('/')->with('success', 'Message created!');
     }
@@ -36,6 +37,8 @@ class MessageController extends Controller
 
     public function update(Request $request, Message $message)
     {
+        $this->authorize('update', $message);
+
         $validated = $request->validate([
             'message' => 'required|string|max:255|min:5',
 
@@ -48,13 +51,18 @@ class MessageController extends Controller
 
     public function edit(Message $message)
     {
+        $this->authorize('update', $message);
+
         return view('messages.edit', compact('message'));
     }
 
     public function destroy(Message $message)
     {
+
+        $this->authorize('delete', $message);
         $message->delete();
         return redirect('/')->with('success', 'Message deleted!');
 
     }
+
 }
